remove inmemory store and manager
