--- graph/diskstorage/keycolumnvalue/KCVSManagerProxy.java
@@ -1,82 +1,0 @@
-/*
- * GRAKN.AI - THE KNOWLEDGE GRAPH
- * Copyright (C) 2019 Grakn Labs Ltd
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package grakn.core.graph.diskstorage.keycolumnvalue;
-
-import com.google.common.base.Preconditions;
-import grakn.core.graph.diskstorage.BackendException;
-import grakn.core.graph.diskstorage.BaseTransactionConfig;
-import grakn.core.graph.diskstorage.StaticBuffer;
-import grakn.core.graph.diskstorage.StoreMetaData;
-
-import java.util.List;
-import java.util.Map;
-
-// This is extended by
-abstract public class KCVSManagerProxy implements KeyColumnValueStoreManager {
-
-    protected final KeyColumnValueStoreManager manager;
-
-    public KCVSManagerProxy(KeyColumnValueStoreManager manager) {
-        this.manager = Preconditions.checkNotNull(manager);
-    }
-
-    @Override
-    public StoreTransaction beginTransaction(BaseTransactionConfig config) throws BackendException {
-        return manager.beginTransaction(config);
-    }
-
-    @Override
-    public void close() throws BackendException {
-        manager.close();
-    }
-
-    @Override
-    public void clearStorage() throws BackendException {
-        manager.clearStorage();
-    }
-
-    @Override
-    public boolean exists() throws BackendException {
-        return manager.exists();
-    }
-
-    @Override
-    public StoreFeatures getFeatures() {
-        return manager.getFeatures();
-    }
-
-    @Override
-    public String getName() {
-        return manager.getName();
-    }
-
-    @Override
-    public List<KeyRange> getLocalKeyPartition() throws BackendException {
-        return manager.getLocalKeyPartition();
-    }
-
-    @Override
-    public KeyColumnValueStore openDatabase(String name, StoreMetaData.Container metaData) throws BackendException {
-        return manager.openDatabase(name, metaData);
-    }
-
-    @Override
-    abstract public void mutateMany(Map<String, Map<StaticBuffer, KCVMutation>> mutations, StoreTransaction txh) throws BackendException;
-
-}
