+++ graph/diskstorage/keycolumnvalue/ttl/TTLKCVSManager.java
@@ -21,20 +21,22 @@ package grakn.core.graph.diskstorage.keycolumnvalue.ttl;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
 import grakn.core.graph.diskstorage.BackendException;
+import grakn.core.graph.diskstorage.BaseTransactionConfig;
 import grakn.core.graph.diskstorage.Entry;
 import grakn.core.graph.diskstorage.EntryMetaData;
 import grakn.core.graph.diskstorage.MetaAnnotatable;
 import grakn.core.graph.diskstorage.StaticBuffer;
 import grakn.core.graph.diskstorage.StoreMetaData;
 import grakn.core.graph.diskstorage.keycolumnvalue.KCVMutation;
-import grakn.core.graph.diskstorage.keycolumnvalue.KCVSManagerProxy;
 import grakn.core.graph.diskstorage.keycolumnvalue.KeyColumnValueStore;
 import grakn.core.graph.diskstorage.keycolumnvalue.KeyColumnValueStoreManager;
+import grakn.core.graph.diskstorage.keycolumnvalue.KeyRange;
 import grakn.core.graph.diskstorage.keycolumnvalue.StandardStoreFeatures;
 import grakn.core.graph.diskstorage.keycolumnvalue.StoreFeatures;
 import grakn.core.graph.diskstorage.keycolumnvalue.StoreTransaction;
 
 import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -45,15 +47,15 @@ import java.util.Map;
  * data with a fixed, CF-wide TTL.  This class stores a fixed TTL set
  * during construction and applies it to every entry written through
  * subsequent mutate/mutateMany calls.
- *
  */
-public class TTLKCVSManager extends KCVSManagerProxy {
+public class TTLKCVSManager implements KeyColumnValueStoreManager {
 
+    private final KeyColumnValueStoreManager manager;
     private final StoreFeatures features;
     private final Map<String, Integer> ttlEnabledStores = Maps.newConcurrentMap();
 
     public TTLKCVSManager(KeyColumnValueStoreManager manager) {
-        super(manager);
+        this.manager = manager;
         Preconditions.checkArgument(manager.getFeatures().hasCellTTL());
         Preconditions.checkArgument(!manager.getFeatures().hasStoreTTL(),
                 "Using TTLKCVSManager with %s is redundant: underlying implementation already supports store-level ttl",
@@ -85,7 +87,7 @@ public class TTLKCVSManager extends KCVSManagerProxy {
     public KeyColumnValueStore openDatabase(String name, StoreMetaData.Container metaData) throws BackendException {
         KeyColumnValueStore store = manager.openDatabase(name);
         int storeTTL = metaData.contains(StoreMetaData.TTL) ? metaData.get(StoreMetaData.TTL) : -1;
-        Preconditions.checkArgument(storeTTL>0,"TTL must be positive: %s", storeTTL);
+        Preconditions.checkArgument(storeTTL > 0, "TTL must be positive: %s", storeTTL);
         ttlEnabledStores.put(name, storeTTL);
         return new TTLKCVS(store, storeTTL);
     }
@@ -93,7 +95,7 @@ public class TTLKCVSManager extends KCVSManagerProxy {
     @Override
     public void mutateMany(Map<String, Map<StaticBuffer, KCVMutation>> mutations, StoreTransaction txh) throws BackendException {
         if (!manager.getFeatures().hasStoreTTL()) {
-            for (Map.Entry<String,Map<StaticBuffer, KCVMutation>> sentry : mutations.entrySet()) {
+            for (Map.Entry<String, Map<StaticBuffer, KCVMutation>> sentry : mutations.entrySet()) {
                 Integer ttl = ttlEnabledStores.get(sentry.getKey());
                 if (null != ttl && 0 < ttl) {
                     for (KCVMutation mut : sentry.getValue().values()) {
@@ -102,13 +104,44 @@ public class TTLKCVSManager extends KCVSManagerProxy {
                 }
             }
         }
-        manager.mutateMany(mutations,txh);
+        manager.mutateMany(mutations, txh);
     }
 
     public static void applyTTL(Collection<Entry> additions, int ttl) {
         for (Entry entry : additions) {
-            ((MetaAnnotatable)entry).setMetaData(EntryMetaData.TTL, ttl);
+            ((MetaAnnotatable) entry).setMetaData(EntryMetaData.TTL, ttl);
         }
+    }
+
+
+    @Override
+    public StoreTransaction beginTransaction(BaseTransactionConfig config) throws BackendException {
+        return manager.beginTransaction(config);
+    }
+
+    @Override
+    public void close() throws BackendException {
+        manager.close();
+    }
+
+    @Override
+    public void clearStorage() throws BackendException {
+        manager.clearStorage();
+    }
+
+    @Override
+    public boolean exists() throws BackendException {
+        return manager.exists();
+    }
+
+    @Override
+    public String getName() {
+        return manager.getName();
+    }
+
+    @Override
+    public List<KeyRange> getLocalKeyPartition() throws BackendException {
+        return manager.getLocalKeyPartition();
     }
